Actual image output!
Currently computing the first batch. All pieces of the project are roughly functional, but the quality of the output can still be improved. There are a number of bugs to fix, and replacing the current spectral decomposition system with diff based clustering will smooth the harsh edges and dramatically increase the accuracy of clusters."